third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnectionIceEvent-constructor-expected.txt

This is a testharness.js-based test.
[FAIL] RTCPeerConnectionIceEvent with no eventInitDict (default)
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] RTCPeerConnectionIceEvent with empty object as eventInitDict (default)
  assert_equals: expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
